feat: improve tests (#26)
* refactor: reorganize test files into core subpackages

* refactor: rename CommandsTest file and update package structure

* refactor(test): moved command tests to separate files

* refactor(test): moved command tests to separate files

* Update src/test/kotlin/hwr/oop/projects/peakpoker/core/deck/DeckTest.kt

---------
